ImageCaptioningAndKeyWordExtraction/requirements.txt
absl-py==0.9.0
astunparse==1.6.3
blis==0.4.1
cachetools==4.1.0
catalogue==1.0.0
certifi==2020.4.5.2
chardet==3.0.4
click==7.1.2
cymem==2.0.3
defusedxml==0.6.0
gast==0.3.3
google-auth==1.16.1
google-auth-oauthlib==0.4.1
google-pasta==0.2.0
grpcio==1.29.0
h5py==2.10.0
idna==2.9
joblib==0.15.1
Keras==2.3.1
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.2
lxml==4.5.1
Markdown==3.2.2
murmurhash==1.0.2
nltk==3.5
numpy==1.18.5
oauthlib==3.1.0
odfpy==1.4.1
opt-einsum==3.2.1
Pillow==7.1.2
plac==1.1.3
preshed==3.0.2
protobuf==3.12.2
pyasn1==0.4.8
pyasn1-modules==0.2.8
PyPDF2==1.26.0
python-docx==0.8.10
PyYAML==5.3.1
regex==2020.6.8
requests==2.23.0
requests-oauthlib==1.3.0
rsa==4.0
scipy==1.4.1
six==1.15.0
spacy==2.2.4
srsly==1.0.2
tensorboard==2.2.2
tensorboard-plugin-wit==1.6.0.post3
tensorflow==2.2.0
tensorflow-estimator==2.2.0
termcolor==1.1.0
thinc==7.4.0
tqdm==4.46.1
urllib3==1.25.9
wasabi==0.6.0
Werkzeug==1.0.1
wrapt==1.12.1
